Changed Shape model class to use double instead of integers for size and position. And added the ability to select and move multiple shapes. And replaced collapsible sidebar functionality with extender control.

+++ src/02350AdvancedDemo.Model/Shape.cs
@@ -16,10 +16,10 @@ namespace _02350AdvancedDemo.Model
         private static int counter = 0;
         public int Number { get; set; } = ++counter;
 
-        public int X { get; set; } = 200;
-        public int Y { get; set; } = 200;
-        public int Width { get; set; } = 100;
-        public int Height { get; set; } = 100;
+        public double X { get; set; } = 200;
+        public double Y { get; set; } = 200;
+        public double Width { get; set; } = 100;
+        public double Height { get; set; } = 100;
 
         public List<string> Data { get; set; }
